split telegram/irc handling to separate modules
